501933 - Changed the script rendering to wrap in a readonly textarea.
As a pre tag, any overly long lines extended the width of the content area indefinitely. On IE, past a certain limit caused problems with the width calculation and the submit button could slide off the right side.

With this change, the script will be held inside of a scrollable textarea, just like when it was originally entered (except on the confirmation page the textarea is readonly). This way, it's rendered just like when the user first entered it, however the size can be held constant and scroll bars are used to navigate long scripts.