506264 - This commit includes:
Fix for iss exporter to extract the ChannelProduct details and send it down to the client.

Fix the importer to not hard code data and extract the exporter content and populate rhnChannelProduct table for all channels instead of jus proxy.

Fix to populate rhnProductName table, product_name_id and channel_product_id in rhnChannel table.
This should help the current issues specific to eus from satsync and UI stand point