506489 - remove the link associated with the org name present in the UI header
Currently, the UI includes information on what organization the user
is logged in to (e.g. Organization: myOrg) and the user can click on
the name to go to the org details page.  While this works for a Satellite
admin, other users do not have permission to access this page; therefore,
removing the link.