Fix mod_jk failure to load.
mod_jk should only be used on RHEL 4, but on Fedora it was trying to be
loaded anyhow. This is due to the fact that if mod_version is not loaded
at all, the IfVersion tag that protects the mod_jk load is ignored and
we load it anyhow. Found that we already load mod_version in
zz-spacewalk-www.conf, but this is parsed much later. Moved the
mod_version load to 00-spacewalk-mod_jk.conf.