508141 - ManageSoftwareChannels - fix broken links for 'delete software channel'
The delete_confirm.pxt has been ported to java; however, there were still
several pages pointing to it.  Updated those.

Also found that the channel id (cid) was not being saved in the request for
ManageSoftwareChannels->channel->Packages; therefore, the delete link on that
page was missing the cid value.  Updated that as well.