Fix API call to remove server custom data value.
Looks like this was being hidden by our recently removed hibernate
exception hiding.

At least in this test, deleting a custom data value from a server was
getting re-saved as the object wasn't actually deleted from it's
association with the server. Seems likely this would be a problem for
actual API use as well.