509027 - kickstart profile edit - update length of supported kernel and post kernel options
The kernel and post kernel options that could be set from the Kickstart edit
profile page would restrict the user to 64 characters for kernel and
post kernel options.  This was too short.  In discussing with others, we
decided to increase this limit to 1024 characters vs removing it entirely.
This will still support a real scenario but avoid adding something entirely
bogus in length.