509376 - add Shared Channels to side navigation of Channels tab
The Shared Channels was added to the Channels tab; however, we missed
including it in the side nav.  This commit adds it.  I also updated the
html/nav navigation, although it doesn't appear to get executed.

Also, changed "Retired channels" to "Retired Channels" to be consistent with
the other channel listings available.