509377 - confirmation pgs for Pkg Install & Remove updated to include pkg arch
The confirmation page that is shown as part of scheduling a package
install or remove for a system was only showing the package NVRE.
This commit will update it to also include the arch (i.e. NVREA) whenever
arch is available.  (This are actions done through Systems->system->
Software->Packages->List/Remove or Install).