510606 - Fix rhn-virtualization package to work with kvm guests. This commit includes fixes for > > - Guest start - We assume pygrub for any guest. This fails for kvm as it the emulates the BIOS loading the first sector of the boot disk and running from there. So we dont need to probe the kernel and ramdisk.
Secondly, the __fixup_config_for_restart checks for the kernel and ramdisk and removes the os block from the xml. This is ok for xen but will miserably fail for kvm and give us Unknown OS Type errors. We should not worry about bootloader or kernel/ramdisk for kvm as its taken care of by the emulator.

With this commit All KVM guest management actions should be functional except reboot. Reboot is not supported yet by libvirt.