Major performance increase for SSM package removal.
With 10k systems this was taking beyond 40 minutes, mostly due to the
use of the per server lookup method, which also does a check for access
after the initial hibernate object lookup. Switched to use batching and
stop doing the unecessary access check, now runs in less than 2 minutes.
Given this is now done asynchronously and not blocking the web UI, this
should be ok.