510329 - Performance enhancements for SSM package upgrade.
After looking at the old perl version, it turns out we never checked which packages can be upgraded on a per sever basis. The functionality was that we'd just flat out schedule upgrades (which really are just installs) for each server selected in the SSM, regardless of whether or not the package was already installed.

This commit changes the Java implementation to use this behavior, drastically increasing the performance in the process.