510785 - Major commit to deal with KS/Distro upgrade scenarios
This commit has a bunch of important changes
1) Concept of a "valid" Distro, implying one
that has the right values for the initrd and kernel path.
An example here will be cases where an external distro was
specified but accidentally got deleted. So the UI and the
cobbler records point to a non existent location..
This fix will put a X in the distros list marking the
distro as invalid. All the profiles using this distro
will also be marked invalid until the distro is either fixed
or the profile's distro is changed.

2) Adds ability to neatly deal with profiles that have
not yet been synced to cobbler. This is typically the be
the upgrade scenario where the user might have the wrong
paths on the distro so they never got created in cobbler
and without a distro no profile can be created which implies
the profiles didn;t themselves sync either. So in this case
the UI will basically  mark the profiles as invalid,
No one will be able to schedule a KS off these profiles.
In addition they will not be able to change any of the properties
of a profile until the distro issue is resolved.
Changes were made in the UI in both the Wizard and the Raw
pages to reflect that.

Also removed several unused old methods...