483606 - Added clause to SSM system retrieval query to filter out proxies.
To be honest, I was a bit surprised to see that the only place this query was used was when changing child channels of systems in the SSM, but that's all there was. So despite it looking like this change would be reflected in a number of places, it only affects channel subscriptions, which is where the proxy filtering matters.