540544 - fix usage of check_user_password
rhnUser.check_user_password check if password is correct
but we want rather call auth_username_password, which return object rhnUser. There is no need for additional checks.
And they even did not worked, becouse we usually have in db crypted
password, but password given as param is not crypted!