Revert "Fix numeric/smallint incompatible types in PostgreSQL."
This reverts commit 9bfcee416f62ec342d249c51b8edc64d57b67ef6.

There are two reasons for the revert:

1) the schema upgrade script won't work. Oracle will throw
"ORA-01440: column to be modified must be empty to decrease precision or scale"
in most of the upgrade situations (i.e. with filled tables)

2) it would be more desirable to change the type of referenced columns
from NUMBER(12) to NUMBER to stay consistent. The upgrade script modifying
NUMBER(12) to NUMBER would work without any problems then.

There's bug #542662 tracking efforts leading to an alternative
solution for the problem addressed by the original commit.