add contributed script apply_errata
This script works against RHN Hosted (the client probably needs a
provisioning entitlement though, haven't checked) and
Satellite/Spacewalk.

The point of the script is to make it easy to apply only particular
errata (i.e. just security errata) but only up to a specific date.
This is useful for people who have to draw a line in the sand
regarding the errata they are allowed to install.