Force correct UTF-8 in changelog where we have history of having broken encoding.
When ISO-8859-1 character like 'ø' is stored in the database without
converting it to UTF-8 (which we now do, see change
5b2705bc2b46f4371844de9d20e3f8610031d630), application

	/network/software/packages/change_log.pxt

was failing with 500 ISE on Spacewalk with Oracle XE. There, the database
has charaster set AL32UTF8, and with NLS_LANG set to english.AL32UTF8,
select from the database would return \x{920f34} (as seen by Perl) and
not \x{fffd}, which is what we get with Oracle XE and NLS_LANG=english.UTF8
or with Oracle EE (which has database character set UTF8).

Thus, we shall try to sanitize the data, before handing it over
to HTML::Entities::encode_entities.