528214 - Encode DBstrings as utf-8 bytes before truncating
If they are left as utf-8 strings and not converted to their
corresponding byte sequence, python treats a multi-byte utf-8
character as one unit.  This causes confusion for things like
splices, especially the one right below this patch; that splice
tries to limit the length of the string to the maximum supported
by the database schema, but fails to do so as it is operating on
a utf-8 string rather than a byte string.

Side note:  NLS_LANG must be set to an appropriate utf-8 locale.