Since there is no unique constraint on rhnChannelComps.channel_id, make sure we always get the latest record.
I know, I know, the application is very careful not to insert
more than one record per channel to that table. But you know what?
I do not trust any of that application code. I do not trust you
people, starting your little sqlplus and inserting an extra record
to rhnChannelComps, just for the sake of doing it. Because people
are sometimes mean. And sometimes they make mistakes.

So why don't we have that constraint there again?