613585 - fixing inter satellite sync and removing HandlerWrap
    HandlerWrap used a static variable which due to certain conditions could cause inter satellite sync
to fail.   in commit 356bddff66b3f7c50ff06f7062d8d111c3f189ff, the satexport module was changed to use
HandlerWrap which introduced the problem.  Since HandlerWrap does not provide much that is useful, it is
being removed and all 3 modules are changed not to use it.  The static variable in question was
HandleWrap.svrHandlers.  It would be set by one request, and in some cases, not re-set by another and
 this caused major issues.