619381 - fixing issue where reprovisioning a system would cause registration to fail in the %post section resulting in the system not being registered at all.
The cause was that we were obtaining a re-activation key regardless of whether or not a reactivation key was already being used within the kickstart(Which would invalidate the one already present).  The little bit of code that i added will go through all activation keys of the kickstart and if one is a reactivation key, will not try to obtain a new one.