593896 - Moved Kickstart Parition UI logic
Order in the Kickstart Partition section in the UI  needed to be preserved.
Due to this made the kickstart partitioning stuff use a blob
instead of a bunch of commands + commandNames. The rendering logic
remains the same except the ordering in the ks partitioning stuff is preserved.
Blob got added to rhnksdata  and the UI some + api command logic got somewhat resturctured
to handle this,