591291 - added new API calls channel.software.mergeErrataWithPackages
the old API calls are preserved.  the new API calls will get the list of
packages that the errata affects, check if they are in the source channel
and if they are, it will then add them to the destination channel.  this should
address Tomas's issues listed in the Bugzilla about pulling in packages
that shouldn't go into the destination channel.