Added constriants on all varchar columns to check '' values.
In short: From now on if you see vn_table_column constraint exception
there is probably need not to insert empty string but NULL value.

Longer description:

We need to track all inserts of empty strings in varchar columns. This
causes problems in our code because Oracle converts them to NULLs but
PosgreSQL does not.

The stored procedure searches for all tables owned by spacewalk user and
containing varchar column and adds constraint named vn_table_column that
checks for empty string. It skips when such a constraint exist.

These constriants will raise error so we will be able to fix all
occurances in a short time. In future we might remove these constriants
when PostgreSQL will be stabilized.