Added Hibernate empty varchar interceptor.
This interceptor checks for empty varchars and issues WARNINGs in the
Catalina.log when such a field is found. It can be also configured to
convert empty varchars to NULLs automatically.

It has been configured to do so.

If you see a warning in form of 'Object XY is setting empty string on'
then you should fix the java code (probably some HTML form) trying to
save empty string in the varchar. This should be changed to NULL.