Prevent satellite-sync from inserting empty strings when it means to insert NULLs.
On Oracle, the empty strings get silently converted to NULLs but on
PostgreSQL they don't, so we'd end up with different data in the
database.

We are glad that vn_rhnchannel_gpg_key_fp reminded us of this problem
in satellite-sync.

Addressing
IntegrityError: new row for relation "rhnchannel" violates check constraint "vn_rhnchannel_gpg_key_fp"