Checkstyle testing report now part of java spec
I have added checkstyle to our spacewalk-java.spec. We wont be able to
build RPM with checkstyle errors.

The build step in our spec is no longer unused. All java sources are
compiled during the build step and the chestyle is processed then.

The jsp precompilation, javadoc and war generation and installation is
taking place in the install step.