Universal function date_diff_in_days introduced
This solves problem with (date2 - date1) diffs. In Oracle we use DATE
columns and in PostgreSQL we use TIMESTAMP WITH TIMEZONE columns. But
the minus operator is not defined for PostgreSQL.

The solution is new universal function date_diff_in_days defined on both
platforms. It substracts the date the very same was as the minus
operator does but with predictible results. It returns number of days
(can be number with fraction digits).

In the code this function can be used on both platforms. Please note the
result can differ a bit on higher fraction digits as the calculation
differs. But the numbers before the digit are always equal.

In PostgreSQL its defined also on date types but this wont be used
probably as we try to use the TIMESTAMP WITH TIMEZONE everywhere.