Sysdate replaced with current_timestamp
In situations WHERE col < sysdate we can replace this with
current_timestamp. This function exists in both Oracle and PostgreSQL
and since there is no arithmetics in this statement the comparsion (or
any other operator) works well.

Under Oracle we use DATE type and in PostgreSQL we use TIMESTAMP WITH
TIMEZONE.