Two classes were not serializabled while putting them into HttpSession
Addressing:

2010-10-29 05:05:57,892 [main] ERROR org.apache.catalina.session.ManagerBase - Exception loading sessions from persistent storage
java.io.WriteAbortedException: writing aborted; java.io.NotSerializableException: com.redhat.rhn.common.util.CharacterMap
    at java.io.ObjectInputStream.readObject0(ObjectInputStream.java:1337)
    at java.io.ObjectInputStream.defaultReadFields(ObjectInputStream.java:1951)
    at java.io.ObjectInputStream.readSerialData(ObjectInputStream.java:1875)
    at java.io.ObjectInputStream.readOrdinaryObject(ObjectInputStream.java:1757)
    at java.io.ObjectInputStream.readObject0(ObjectInputStream.java:1333)
    at java.io.ObjectInputStream.readObject(ObjectInputStream.java:352)
    at org.apache.catalina.session.StandardSession.readObject(StandardSession.java:1402)
    at org.apache.catalina.session.StandardSession.readObjectData(StandardSession.java:931)
    at org.apache.catalina.session.StandardManager.doLoad(StandardManager.java:394)
    at org.apache.catalina.session.StandardManager.load(StandardManager.java:321)

We put many objects into the Http Session. The java web container is able to
(undeterministically) store (seliazilable) the session on the disc. Typically
when the session occupies lot of memory.

According to the J2EE Servlet Spec we should mark *all* classes that are being
serialized as Serializable and check it. I have addressed the issue above and also
went through all our java source code and fixed one more object that should be
Serializable.

The problem could be in the Elaborator class because its a Map of objects. We should
keep an eye on objects being inserted in this interface - all need to be Selializable.