require versions which can handle cache_only
this older relases does not understood cache_only parameter for actions
if we call them with cache_only, we will create traceback.
It is not possible to require new version directly, because user do not
need this packages and can (and usually really do) operate without them.
So I choose to use conflict to not allow user have new rhn-client-tools
and old package with action class.