644239 - do not check minor version of xml_dump_version
All dumps with the same major version are (and in future should) be backward
and forward compatible.
Old satellites can understoo new dumps as new xml attributes and elements, which old
satellite do not understood are ignored (like sha256 checksums).
And new satellites understood older dumps. We always have to be able to read old dumps.

If we ever move to new incompatible format, we will have to bump up major version of xml_dump.