656241 - require libvirt
Originaly libvirt provided libvirt.so.0 (which is required by libvirt-python,
but in RHEL6 it is now separated to libvirt-client.
Therefore we have to require libvirt extra, to get libvirtd service.
But there is problem with RHEV. On RHEL6, it use libvirt - thats good.
But on RHEL5 it conflict with libvirt and use vdsm and bunch of other packages.
I have no clue how to distinguish regular Fedora/RHEL and RHEV. :(
So on RHEL5 we will keep status quo and behave as we behaved.