677549 - do not require spacewalk-backend-libs in the same version
we should require only the minimal version ve really need
last incompatible move of code was (according to my search)
done in commit 4c3b684480245f9c6ec6001158c3cbde38243dd8
Therefore we should be happy with anything greater then 0.8.28

On the other hand last incompatible move of code from libs POV was
41d7eb19fe69de34821f6ef58a86002d12091504, so we should conflict with
< 0.8.28-1

I'm bit of lazy with that Conflict. Correctly it should conflict
with subset of submodules of spacewalk-backend. But the result should be same.