rhn-satellite-exporter with start-date and end-date
This patch fix problem in rhn-satellite-exporter on PostgreSQL:
When we are trying to export dump with options --start-date and/or
--end-date with PostgreSQL backend these dates have only day precision, but
with Oracle backend they have second precision.

On PostgreSQL DATE is represented as day only and the time information is
striped. We want to represent these dates as TIMESTAMP so we can preserve
time and have second precision.