679353 - automatically detect system re-registration
This fix is to avoid the need to re-start osad service after
the machine has been re-registered w/ Spacewalk.

By setting the select() timeout (currently to 180 seconds), we
will allow osad to re-enter the process_once() routine, which
takes care of the re-registration detection. Otherwise the system
could have been stuck in an infinite select().