fix vm-state poller
there is a bug in spacewalk, that the state of your virtual guests is not
shown correctly if all machines are in state "stopped".
As long as at least one guest is running, the indicatior on the webpage:

Systems->[MACHINE]->Virtualization->Details

is correct but if you turn off the last machine, the state will still be
"running" for that guest. Spacewalk will never notice the shutdown.

That's because of a bug in poller.py on the client (virtual host) side. There
is an optimization to exit the program if the domain_list from
poll_hypervisor() is empty.
That list is empty when all virtual guests are stopped, so if you shutdown the
last running virtual guest, the poller.py will exit early and will return an
empty result to the Spacewalk server who thinks now that nothing has happened.