Copying kopts to the xen distro.
I noticed that the XEN distro which is automatically created if
possible, when you add a kickstartable tree, does not inherit the
kernel/post-kernel parameters from the original distro.
So the user has no chance to change the kopts for that distro, except
with the cobbler CLI.

Here is a simple patch that copies the kopts to the xen distro.