Insert pxt session errors on postgres
Hello,

We found a problem with an sql statement not working on postgresql, causing
errors when trying to browse perl pages while not logged in to spacewalk.

To reproduce, try to navigate to one of these pages when not logged in on a
postgres based installation:

https://<hostname>/help/copyright.pxt
https://<hostname>/help/about.pxt

Attached please find a patch for the issue.

Regards,
Johannes

--
SUSE LINUX Products GmbH, HRB 16746 (AG Nürnberg)
GF: Jeff Hawn, Jennifer Guild, Felix Imendörffer

>From 76dfe76b7cb8a7882951849ed537a66154aa86cc Mon Sep 17 00:00:00 2001
From: Johannes Renner <jrenner@suse.de>
Date: Wed, 15 Aug 2012 15:23:59 +0200
Subject: [PATCH] Fix insert pxt session for postgres when not logged in