Expose extra packages / systems with extra packages
Extra packages in this context mean packages installed on a registered
system which are not present in any of the system's subscribed channel
(with the option to remove these packages).

This change adds the following pages:
* systems/ExtraPackagesSystems
* systems/details/packages/ExtraPackagesList
* systems/details/packages/RemoveExtraConfirm

and the following two API calls:
* system.listSystemsWithExtraPackages()
* system.listExtraPackages()

This is a re-work of a code originally found in SUSE Manager 1.7.