spacecmd: don't exit when invalid arguments are passed to a function
We use optparse to handle arguments to spacecmd functions.  By default,
it will exit when there is an invalid argument passed to it.  The only
way to override this is to sublcass OptionParser and change the behavior
of error().  Now, we just raise an exception with the original error,
which will be printed for the user to read.