use the server timezone as the default for the first user
Do not use America/New_York as the default timezone for the first new user (subsequent users take as default the one from the user creating the new user), but take the one configured for the server using java.util.TimeZone.getDefaultTimeZone(). As the rhnTimeZone table does not contain all timezones, but one per offset, we take the first one that matches the offset.