A lot of JUnit cleanup
A number of tests were committing changes to the DB and not cleaning up
after themselves.  In addition to leaving the spacewalk DB a mess, this
consumes entitlements - repeated executions eventually (or even immediately!)
runs the instance out of entitlements, and further tests (and further use)
of the spacewalk instance fails.

* Make sure user/orgs created can be traced back to their creators
* Make sure that tests that issue commits, explicitly clean up