There is an issue with the default yum repo label
There is an issue with the default yum repo label with some of the new repositories, particularly the Scientific Linux and Oracle Linux Spacewalk Server/Client repos. The full text expands to greater than 64 chars, which is the limit of the database field that stores the repository name. If you try to add these channels, you get an error because the DB rejects the change.

This patch changes the default prepend text to "External - " which shortens the text sufficiently so that all the channels can be created. The alternative is to specify yum_repo_label for each repository that may go over 64 chars. Not sure which solution you prefer, but this seemed simpler and it would at least be consistent for all repos.