DataSourceParserTest: fix Oracle case-depentent tests
Username queries on the all_catalog table on Oracle can return results in the
wrong case (in particular: uppercase). Test code assumes that usernames are
lowercase in config files (see DataSourceParser.java:96), so those tests will
fail if case does not match, and this was actually verified with a 11g XE
instance.

This patch adds LOWERCASE() calls to restore correct results.