spacecmd: Fix session validation
Turns out that although spacewalk (Red Hat Satellite 5.5) takes a
session argument for the api.getApiNamespaces() API call, this session is not
validated, i.e. the call will succeed no matter if the session is
valid or not. Which makes this particular call quite a bad choice
validating the session... Trying now to use user.listAssignableRoles
instead