917070 - catch jabberd connection errors
This fix is to avoid an ugly traceback which is printed into the console
when osa-dispatcher starts while jabberd is down. With this fix, even
when jabberd is down, osa-dispatcher will print an error message, yet
the service will report a starting success -- not to break osad
and previously expected behavior. After jabberd starts, osa-dispatcher
will eventually connect to it.