Added Oracle Linux channels for UEKR3, as well as Spacewalk 2.0 Server/Client for OL6 and Client for OL5
Oracle has shipped UEK Release 3 as well as Oracle built and signed versions of Spacewalk 2.0 Server (for Oracle Linux 6-x86_64) and Spacewalk 2.0 Client for (OL5/OL6). This patch adds those channels to spacewalk-common-channels.ini.

I specified the Spacewalk channels as children of the parent OL5/OL6 channels as they inherit the GPG keys from the parent. They also use an alternative naming so as not to conflict with anyone who wants to use the upstream server/client channels.