1040540 - have package search return all matching results
This fix is to make sure:
1. Search results as we get them from the search server will be split
into chunks of 1000, since we need to avoid ORA-01795 errors when
we select further info for these packages from the database.

2. 'max_hits_returned' config option is no longer needed, and will
be only taken into account when it's specified in rhn.conf. That is,
by default, the tomcat application won't be trimming the search
results according to 'max_hits_returned', not unless the config
option is present in rhn.conf

Both of the above will lead to all search matches being present
in Spacewalk webui.