Introduce a date-time picker.
The picker can be used as a tag:

<rhn:datepicker data="${picker}"/>
Where picker is the DatePicker java object currently used.

The date-picker fragment jsp is kept for backwards compatibility, but now it
is only one line that uses the jsp tag.

It generates backward compatibility input tags with date_hour, date_minutes, date_am_pm...,
using Javascript to be backwards compatible with the old tag, so it should work in all pages.

The date is displayed in a localized format when the calendar is not open.
The calendar is localized using the month names from the DatePicker class and related classes.