/rhn/admin/config/MonitoringConfig.do help texts ((e.g. 'redhat.com' for myemail@…)) should be bellow input bars (see /rhn/activationkeys/Edit.do?tid=1 how it should look like)
This form is generated from the backend by iterating a config description and a config
value map. The description is already concatenated with the help.
(This was already like that in Spacewalk 2.0 so this issue is not Bootstrap-port
specific)

We use a bit of page-specific Javascript to detect text of the type
"This is something (with some help)" to move the help and apply
the right style to it.
If the backend can provide the text separated then we should
try that instead.

(cherry picked from commit 12b33028b0fb4ee0662b45dd6decadefa73959a8)