Form names are only available as name attributes now, not ids.
I think this is caused by 51c3e001a7755d016fe4ad40f048370927017f79.
I think that when struts-taglib renders html:form elements as
html it sticks the name in a "name" attribute, and when rendering
as xhtml it sticks the name in an "id" attribute. But the documentation
is sparse, so it's hard to tell.