A couple of mode queries didn't quite fit into ChannelTreeNode objects.
This is because they were doing things the old Perlish viewy way instead
of the new Javaish objecty way. If we want to re-write owned_channels_tree
then we can drop the rhnChannelTreeView entirely, but for now I'll
just modify the data returned to make it fit in a ChannelTreeNode.