Fix Debian port of rhnreg_ks to work also with more recent Debian versions
Recent Debian versions may include the architecture in the version string used
in the package file name. This is used in multi-arch setups.

This patch was submitted by Carsten Menzel <carsten.menzel@bechtle.com> as fix
for Debian Bug #703582 (http://bugs.debian.org/703582). I tested it and can
confirm that it fixes the issue.