Set cobbler hostname variable when calling system.createSystemRecord
When adding a record for a non-registered system, there is no way to change the
hostname variable which is used in the autoinstallation profiles.
So we set it from the server hostname or name, depending which one is available.

(cherry picked from commit ba8ca0309ece2fb81d7cca9c6e8b2364c3b90a99)