migrate clone channel page from perl -> java
This is a large change. It also introduces a fast (hibernate-free) way
to clone errata, and significantly speeds up the hibernate errata clone
helper. I'm not changing the nav urls to point to the java clone page
yet because we need to port the errata clone page to java so that the
'clone and select errata' option has a place to redirect you to.