Improve osad's handling of the rhn_check process.
osad did a waitpid during the _message_callback, which resulted in any message
received while rhn_check is running accumulated. Once the event-loop continues, this
resulted in several calls to rhn_check back to the server.

Now we use a subprocess handle and continue the event loop. The interval is reduced
when rhn_check is executing in order to monitor the process. While it is running
checking messages from dispatcher are processed but rhn_check is not run again.

If rhn_check fails, the event loop keeps calling it, in order to exaust the server-side
remaining_tries, or hope that a future call works.