Update 021-rhnServerNetwork-trigger.sql.oracle
This permanently fixes https://bugzilla.redhat.com/show_bug.cgi?id=1215151 in a proper way. 1. it does not "IF EXIST" the trigger, since that does not exist in oracle. 2. it does indeed check if the trigger exists, and if so, removes it, otherwise just continues, as the trigger does not exist anymore.