Unify profile creation/update with one submit button instead of two.
This removes the Upload File button on the Create/Update Kickstart Profile page
which does the same thing as the Create or Update button but uses the file input
instead of the textbox. With this patch the Create and Update button will use
either the file input or the textbox depending on which one is provided. If both
are provided it will show an error telling the user to only provide one
of the two.