Make arch x86_64 the default when creating new channels.
As we set the parent channel architecture using Javascript with the compatibility table,
there is no need to initially set the architecture of the parent channel as the selected
one, because initially the parent channel is None.

So, set x86_84 as the initially selected one, and then any change is handled using
the compatibility table on the client side.