rename ChannelProduct#beta to ChannelProduct#betaMarker
There is already an (innapropiately, as it is a marker or flag)
named "beta" property, that has getBeta() and setBeta(str).

Problem is that there is also a boolean helper, which is the right way of
accessing it from Java code, but isBeta clashes with getBeta/setBeta
having a completely different type.

In Java 7 Class.getDeclaredMethods() does not guarantee order anymore.
http://bugs.java.com/bugdatabase/view_bug.do?bug_id=7023180

So running tests with OpenJDK, Hibernate will get wrong method for the property
"beta" and the following error will happen:

java.lang.ClassCastException: java.lang.Boolean cannot be cast to java.lang.String